+++ example/DLRA_ginzburg_landau/ginzburg_landau_tests_riccati.f90
@@ -171,7 +171,7 @@ contains
       character(len=256)                           :: fbase, fchomp
       real(dp)                                     :: X_out(N,N)
       logical                                      :: exist_file
-      real(wp),                      allocatable   :: meta(:)
+      real(dp),                      allocatable   :: meta(:)
       character(len=128)                           :: note
       ! timing
       integer                                      :: clock_rate, clock_start, clock_stop
@@ -297,7 +297,7 @@ contains
       real(dp)                                     :: X_out(N,N)
       logical                                      :: exist_file
       character(len=128)                           :: note
-      real(wp),                      allocatable   :: meta(:)
+      real(dp),                      allocatable   :: meta(:)
       ! timing
       integer                                      :: clock_rate, clock_start, clock_stop
